+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Polyverse Boost Automatic Application Modernization
 
 ### Enhancements
 - Support excluding files and folders from Analysis by right-click "Exclude..." context-menu in File Explorer (writes exclusions to .boostignore file)
+- Initial Blueprinting of your software project will automatically generate an analysis exclusion list (aka .boostignore) file based on the files in your project and your project type
 - Analysis Errors are now loaded immediately on Project/Folder Load, without manually opening the Notebook with the errors
 - Source-level Errors are loaded in Diagnostic Collections "Problems" for each area - e.g. "Security", "Data Compliance", "Performance"